% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoAdjust.R
\name{AutoAdjust}
\alias{AutoAdjust}
\title{AutoAdjust}
\usage{
AutoAdjust(meshP, meshT, edge, centroid, AspectR, AR)
}
\arguments{
\item{meshP}{Matrix (2 x n) containing coordinate points of the mesh nodes.}

\item{meshT}{Matrix (3 x n) containing the number of the coordinate point that forms a given triangle within the mesh.}

\item{edge}{Coordinate points of the initial geometry.}

\item{centroid}{Matrix (2 x n) of triangle elements.}

\item{AspectR}{Aspect ratio of each triangle element.}

\item{AR}{minimum desired aspect ratio, numeric value.}
}
\value{
Generates new mesh and centroid tables
\item{Meshpts}{Includes both new mesh coordinate points and triangulation of points.}
\item{Centroids}{Centroid positions for each triangle element.}
}
\description{
Allows for automatic refinement of the triangular mesh generated based on given parameters. Will remove elements that are outside the margin of the geometry.
}
\examples{
# meshP = MeshPts$p #mesh points
# meshT = MeshPts$T #triangulation list
# edge =  Line #original geometry
# centroid = centroids
# AspectR = AspectRatio #dimensional value
# AR = 10 #aspect ratio limit for elimination
# AutoAdjust(meshP, meshT, edge, centroid, AspectR, AR)

}
