% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamBC.R
\docType{data}
\name{beamBC}
\alias{beamBC}
\title{Boundary conditions applied to each node. Obtained from function: beamApplyBC}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 8 rows and 1 columns.
}
\usage{
beamBC
}
\description{
Boundary conditions applied to each node. Obtained from function: beamApplyBC
}
\keyword{datasets}
