% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamDime.R
\docType{data}
\name{beamDime}
\alias{beamDime}
\title{Dimensional data for beam elements. Includes area, length, aspect ratio, angles and lengths of elements. Obtained from function: beamDimensions}
\format{
An object of class \code{list} of length 7.
}
\usage{
beamDime
}
\description{
Dimensional data for beam elements. Includes area, length, aspect ratio, angles and lengths of elements. Obtained from function: beamDimensions
}
\keyword{datasets}
