% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamEmat.R
\docType{data}
\name{beamEmat}
\alias{beamEmat}
\title{List of element matrices for each element. Obtained from function: beamElementMat}
\format{
An object of class \code{list} of length 3.
}
\usage{
beamEmat
}
\description{
List of element matrices for each element. Obtained from function: beamElementMat
}
\keyword{datasets}
