% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamExMat.R
\docType{data}
\name{beamExMat}
\alias{beamExMat}
\title{List of element matrices for each element. Obtained from function: beamElementMat}
\format{
An object of class \code{list} of length 3.
}
\usage{
beamExMat
}
\description{
List of element matrices for each element. Obtained from function: beamElementMat
}
\keyword{datasets}
