% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamFV.R
\docType{data}
\name{beamFV}
\alias{beamFV}
\title{Load vector produced from function function: beamForceVector}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 5 rows and 1 columns.
}
\usage{
beamFV
}
\description{
Load vector produced from function function: beamForceVector
}
\keyword{datasets}
