% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamGeo.R
\docType{data}
\name{beamGeo}
\alias{beamGeo}
\title{Sample geometry for beam. Includes shape, discretization table, boundary conditions, thickness, and material details.}
\format{
An object of class \code{list} of length 12.
}
\usage{
beamGeo
}
\description{
Sample geometry for beam. Includes shape, discretization table, boundary conditions, thickness, and material details.
}
\keyword{datasets}
