% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamGLforce.R
\docType{data}
\name{beamGLforce}
\alias{beamGLforce}
\title{Global and Local loading force matrices obtained from function: beamGLForces}
\format{
An object of class \code{list} of length 2.
}
\usage{
beamGLforce
}
\description{
Global and Local loading force matrices obtained from function: beamGLForces
}
\keyword{datasets}
