% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamGlobalEM.R
\name{beamGlobalEM}
\alias{beamGlobalEM}
\title{beamGlobalEM}
\usage{
beamGlobalEM(beamExEM)
}
\arguments{
\item{beamExEM}{Expanded Element Matrix}
}
\value{
Produces large (n x n) global matrix
\item{GlobalMat}{Global matrix}
}
\description{
Generates global stiffness matrix for beams.
}
\examples{
data(beamExMat)

beamExEM = beamExMat
GMat = beamGlobalEM(beamExEM)

}
