% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamGloMat.R
\docType{data}
\name{beamGloMat}
\alias{beamGloMat}
\title{Global element matrix, obtained from function: beamGlobalEM}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 8 rows and 8 columns.
}
\usage{
beamGloMat
}
\description{
Global element matrix, obtained from function: beamGlobalEM
}
\keyword{datasets}
