% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamND.R
\docType{data}
\name{beamND}
\alias{beamND}
\title{Global nodal displacement, obtained from function: beamNodeDis}
\format{
An object of class \code{list} of length 3.
}
\usage{
beamND
}
\description{
Global nodal displacement, obtained from function: beamNodeDis
}
\keyword{datasets}
