% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamREM.R
\docType{data}
\name{beamREM}
\alias{beamREM}
\title{Reduced element matrix calculated from the expanded element matrix. Obtained from function: beamReducedEM}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 5 rows and 5 columns.
}
\usage{
beamREM
}
\description{
Reduced element matrix calculated from the expanded element matrix. Obtained from function: beamReducedEM
}
\keyword{datasets}
