% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamStressResult.R
\docType{data}
\name{beamStressResult}
\alias{beamStressResult}
\title{FEA results for the beam model. Obtained from function: beamStress}
\format{
An object of class \code{numeric} of length 3.
}
\usage{
beamStressResult
}
\description{
FEA results for the beam model. Obtained from function: beamStress
}
\keyword{datasets}
