% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamsUDL.R
\docType{data}
\name{beamsUDL}
\alias{beamsUDL}
\title{Uniformly distributed load on beam surface}
\format{
An object of class \code{list} of length 3.
}
\usage{
beamsUDL
}
\description{
Uniformly distributed load on beam surface
}
\keyword{datasets}
