% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bound.R
\docType{data}
\name{bound}
\alias{bound}
\title{Boundary conditions applied to each node. Obtained from function: ApplyBC}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 100 rows and 1 columns.
}
\usage{
bound
}
\description{
Boundary conditions applied to each node. Obtained from function: ApplyBC
}
\keyword{datasets}
