% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cart.R
\docType{data}
\name{Cart}
\alias{Cart}
\title{Sample geometry. Matrix with x and y coordinates for initial shape.}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 11 rows and 2 columns.
}
\usage{
Cart
}
\description{
Sample geometry. Matrix with x and y coordinates for initial shape.
}
\keyword{datasets}
