% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanpoly.R
\docType{data}
\name{cleanpoly}
\alias{cleanpoly}
\title{Cleaned nodal distribution in and on polygon. Obtained from function: Threshpts}
\format{
An object of class \code{list} of length 2.
}
\usage{
cleanpoly
}
\description{
Cleaned nodal distribution in and on polygon. Obtained from function: Threshpts
}
\keyword{datasets}
