% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dime.R
\docType{data}
\name{dime}
\alias{dime}
\title{Dimensional data for mesh elements. Includes area, length, aspect ratio, angles and lengths of elements. Obtained from function: Dimensions}
\format{
An object of class \code{list} of length 6.
}
\usage{
dime
}
\description{
Dimensional data for mesh elements. Includes area, length, aspect ratio, angles and lengths of elements. Obtained from function: Dimensions
}
\keyword{datasets}
