% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displacN.R
\docType{data}
\name{displacN}
\alias{displacN}
\title{Global nodal displacement, obtained from function: NodeDis}
\format{
An object of class \code{list} of length 2.
}
\usage{
displacN
}
\description{
Global nodal displacement, obtained from function: NodeDis
}
\keyword{datasets}
