% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ElementMat.R
\name{ElementMat}
\alias{ElementMat}
\title{ElementMat}
\usage{
ElementMat(meshP, meshT, Nu, Y, Thick)
}
\arguments{
\item{meshP}{Matrix (2 x n) containing coordinate points of the mesh nodes.}

\item{meshT}{Matrix (3 x n) containing the number of the coordinate point that forms a given triangle within the mesh.}

\item{Nu}{Value of Poisson's ratio}

\item{Y}{Value of Young's (Elastic) modulus}

\item{Thick}{Value of the thickness of the mesh, a positive value must be given.}
}
\value{
Generates initial element matrix needed for the finite element model.
\item{EMPStress}{An element matrix of the geometry under stress.}
\item{EMPStrain}{An element matrix of the geometry under strain.}
}
\description{
Generates an element stiffness matrix
}
\examples{
# Y = 20 #Elastic modulus
# Nu = 0.45 #Poisson ratio
# Thick = 0.001
# DOF = 6
# meshP = MeshPts$p
# meshT = MeshPts$T
# test7 = ElementMat(meshP, meshT, Nu, Y, Thick)

}
