% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expSurf.R
\docType{data}
\name{expSurf}
\alias{expSurf}
\title{Expanded element matrix for surface forces. Obtained from function: ExpandSFT}
\format{
An object of class \code{list} of length 50.
}
\usage{
expSurf
}
\description{
Expanded element matrix for surface forces. Obtained from function: ExpandSFT
}
\keyword{datasets}
