% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fea_EM.R
\docType{data}
\name{fea_EM}
\alias{fea_EM}
\title{List of element matrices for each element. Obtained from function: ElementMat}
\format{
An object of class \code{list} of length 2.
}
\usage{
fea_EM
}
\description{
List of element matrices for each element. Obtained from function: ElementMat
}
\keyword{datasets}
