% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fea_ExEM.R
\docType{data}
\name{fea_ExEM}
\alias{fea_ExEM}
\title{List of large expanded element matrices calculated from the element matrix. Obtained from function: ExpandEM}
\format{
An object of class \code{list} of length 78.
}
\usage{
fea_ExEM
}
\description{
List of large expanded element matrices calculated from the element matrix. Obtained from function: ExpandEM
}
\keyword{datasets}
