% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fea_result.R
\docType{data}
\name{fea_result}
\alias{fea_result}
\title{FEA results. Produces list with results from local stresses including Stress, Strain, and Stress from Strain. Obtained from function: LocalStress}
\format{
An object of class \code{list} of length 3.
}
\usage{
fea_result
}
\description{
FEA results. Produces list with results from local stresses including Stress, Strain, and Stress from Strain. Obtained from function: LocalStress
}
\keyword{datasets}
