% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompleteFEMStrain.R
\name{FEMStrain}
\alias{FEMStrain}
\title{FEMStrain}
\usage{
FEMStrain(meshP, meshT, centroid, BoundConx, BoundCony, SFShear,
SFTensile, Length, area, Fx, Fy, Y, Nu, Thick)
}
\arguments{
\item{meshP}{Matrix (2 x n) containing coordinate points of the mesh nodes.}

\item{meshT}{Matrix (3 x n) containing the number of the coordinate point that forms a given triangle within the mesh.}

\item{centroid}{Matrix (2 x n) containing coordinate points of the centroid of each triangular element.}

\item{BoundConx}{Boundary constraint for nodes in the x-direction}

\item{BoundCony}{Boundary constraint for nodes in the y-direction}

\item{SFShear}{Magnitude of positive shear traction; if there is no surface traction then SFShear = 0}

\item{SFTensile}{Magnitude of tensile surface traction; if there is no surface traction then SFTensile = 0}

\item{Length}{Truss length}

\item{area}{Triangle element area}

\item{Fx}{Load vector for the x-direction}

\item{Fy}{Load vector for the y-direction}

\item{Y}{Value of Young's (Elastic) modulus}

\item{Nu}{Value of Poisson's ratio}

\item{Thick}{Value of the thickness of the mesh, a value must be given.}
}
\value{
Completes the FEM to generate values of stress and strain; produces three (3) [3 x n] matrices.
\item{Strain}{Calculated strain. [x, y, tau]}
\item{Stress}{Calculated stress in pascals. [x, y, tau]}
\item{StressStrain}{Stress as calucated from strain. [x, y, tau]}
}
\description{
Creates a complete finite element model using strain for a given 2D mesh under specified boundary conditions (constrain and load).
}
\examples{
Y = 17e9
Nu = 0.45
Thick = 0.001
DOF = 6

# meshP = MeshPts$p
# meshT = MeshPts$T
# centroid = Centroids
# BoundConx = numeric(NROW(meshP))
# BoundCony = numeric(NROW(meshP))
# BoundConx[1:NROW(meshP)] = BoundCony[1:NROW(meshP)] = 1
# SFShear = 0
# SFTensile = 0
# Length = TrussLength
# area = Area
# Fx = 20
# Fy = 10
# FEAtest = FEAStrain(meshP, meshT, centroid, BoundConx, BoundCony, SFShear, SFTensile, Length, area, Fx, Fy, Nu, Y, Thick)

# Following the FEAtest, plot for value map:
# PlotVal = FEAtest$Strain
# Kol = 3
# a= 1; b= 2; c=3; d=4; e=5; f=6; g=7; h=8; i=9
# PlotSystem(meshP, meshT, PlotVal, Kol, a, b, c, d, e, f, g, h, i)

}
