% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glfor.R
\docType{data}
\name{glfor}
\alias{glfor}
\title{Global and Local loading force matrices obtained from function: GLForces}
\format{
An object of class \code{list} of length 2.
}
\usage{
glfor
}
\description{
Global and Local loading force matrices obtained from function: GLForces
}
\keyword{datasets}
