% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLForces.R
\name{GLForces}
\alias{GLForces}
\title{GLForces}
\usage{
GLForces(meshP, meshT, GMat, GlobalND, EMatrixlist)
}
\arguments{
\item{meshP}{Matrix (2 x n) containing coordinate points of the mesh nodes.}

\item{meshT}{Matrix (3 x n) containing the number of the coordinate point that forms a given triangle within the mesh.}

\item{GMat}{Global matrix}

\item{GlobalND}{Global nodal displacement}

\item{EMatrixlist}{Element matrix list}
}
\value{
Matrices of global and local forces
\item{GForce}{Large global force matrix.}
\item{Lforce}{Large local force matrix.}
}
\description{
Uses nodal displacements to determine global and local forces at each node
}
\examples{
# meshP = MeshPts$p
# meshT = MeshPts$T
# GMat = Gm #global matrix
# GlobalND = GlobalND #global nodal displacement
# EMatrixlist = EMPStress
# GLForces(GMat, GlobalND, EMatrixlist, meshT)

}
