% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gloMat.R
\docType{data}
\name{gloMat}
\alias{gloMat}
\title{Global element matrix, obtained from function: GlobalMat}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 100 rows and 100 columns.
}
\usage{
gloMat
}
\description{
Global element matrix, obtained from function: GlobalMat
}
\keyword{datasets}
