% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\docType{data}
\name{load}
\alias{load}
\title{Load vector produced from function function: ForceVector}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 94 rows and 1 columns.
}
\usage{
load
}
\description{
Load vector produced from function function: ForceVector
}
\keyword{datasets}
