% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSystem.R
\name{PlotSystem}
\alias{PlotSystem}
\title{PlotSystem}
\usage{
PlotSystem(meshP, meshT, PlotVal, Kol, a, b, c, d, e, f, g, h, i)
}
\arguments{
\item{meshP}{Matrix (2 x n) containing coordinate points of the mesh nodes.}

\item{meshT}{Matrix (3 x n) containing the number of the coordinate point that forms a given triangle within the mesh.}

\item{PlotVal}{Value to be plotted, either stress or strain, return from function LocalStress function.}

\item{Kol}{Which plot value column to plot (choose 1, 2, or 3)? 1 = X-stress/strain; 2 = Y-stress/strain; 3 = Shear stress/strain.}

\item{a}{Threshold for royal blue}

\item{b}{Threshold for sky blue}

\item{c}{Threshold for bluegreen}

\item{d}{Threshold for dark green}

\item{e}{Threshold for green}

\item{f}{Threshold for gold}

\item{g}{Threshold for orange}

\item{h}{Threshold for pink}

\item{i}{Threshold for red}
}
\value{
Plot of colored polygon with mesh colored based on the plot value
}
\description{
Generates heat map for given stress or strain on the geometry. Threshold values for the color must be assigned.
}
\examples{
# meshP = MeshPts$p #mesh points
# meshT = MeshPts$T #triangulation list
# PlotVal = FEAtest$Stress  #output value wanted in plot
# Kol =  3 #plotting shear stress/strain
# a= 1; b= 2; c=3; d=4; e=5; f=6; g=7; h=8; i=9 #color threshold values from min to max
# PlotSystem(meshP, meshT, PlotVal, Kol, a, b, c, d, e, f, g, h, i)

}
