% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyshape.R
\docType{data}
\name{polyshape}
\alias{polyshape}
\title{Sample geometry converted into a 2D polygon. Polygon data that specifies all coordinate, coordinates that are within the geometry and coordinates that construct the lines of the geometry. Obtained from function: SinglePoly}
\format{
An object of class \code{list} of length 3.
}
\usage{
polyshape
}
\description{
Sample geometry converted into a 2D polygon. Polygon data that specifies all coordinate, coordinates that are within the geometry and coordinates that construct the lines of the geometry. Obtained from function: SinglePoly
}
\keyword{datasets}
