% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduc_EM.R
\docType{data}
\name{reduc_EM}
\alias{reduc_EM}
\title{Reduced element matrix calculated from the expanded element matrix. Obtained from function: ReducedEM}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 94 rows and 94 columns.
}
\usage{
reduc_EM
}
\description{
Reduced element matrix calculated from the expanded element matrix. Obtained from function: ReducedEM
}
\keyword{datasets}
