% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduc_SF.R
\docType{data}
\name{reduc_SF}
\alias{reduc_SF}
\title{Reduced surface force matrix calculated from expanded element matrix. Obtained from function: ReducedSF}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 100 rows and 1 columns.
}
\usage{
reduc_SF
}
\description{
Reduced surface force matrix calculated from expanded element matrix. Obtained from function: ReducedSF
}
\keyword{datasets}
