% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReducedEM.2d.R
\name{ReducedEM.2d}
\alias{ReducedEM.2d}
\title{ReducedEM.2d}
\usage{
ReducedEM.2d(GMat, NodeKnownL)
}
\arguments{
\item{GMat}{Global stiffness matrix}

\item{NodeKnownL}{data frame with constraint parameters applied to each node in the x and y directions. Formatted for use in reduced element matrix. Generated from ApplyBC function.}
}
\value{
Produces a large matrix.
\item{ReducedEM}{Reduced element matrix.}
}
\description{
Reduced stiffness matrix - use boundary condition to reduce matrix to smaller form by removing systems that are bound.
}
\examples{
data(gloMat)
data(bound)
GMat = gloMat
NodeKnownL = bound
reduc_EM = ReducedEM.2d(GMat, NodeKnownL)

}
