% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SinglePoly.R
\name{SinglePoly}
\alias{SinglePoly}
\title{SinglePoly}
\usage{
SinglePoly(x, y, ptDS, ptDL)
}
\arguments{
\item{x}{X-coordinates for geometry.}

\item{y}{Y-coordinates for geometry.}

\item{ptDS}{Density of points desired within the geometry.}

\item{ptDL}{Density of points desired at the perimeter of the geometry.}
}
\value{
Coordinate points of nodes distributed within and on the line of a given geometry.
\item{AllCoords}{all coordinate points distributed across the geometry.}
\item{Within}{all coordinate points within the geometry ONLY.}
\item{Line}{all coordinate points that lay on the perimeter of the geometry ONLY.}
}
\description{
Generates a mesh for polygon with a single continuous geometry
}
\examples{
x= c(0.23, 0.61, 0.75, 0.61, 0.23, -0.23, -0.61, -0.75, -0.61, -0.23, -0.23)
y= c(0.62, 0.38, 0.51, -0.38, -0.62, -0.62, -0.38, -0.51, 0.38, 0.62, 0.65)
ptDS = 30
ptDL = 20
SinglePoly(x, y, ptDS, ptDL)

}
