% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurfTrac.R
\docType{data}
\name{SurfTrac}
\alias{SurfTrac}
\title{List of element matrices with surface traction. Obtained from function: SurfaceTraction}
\format{
An object of class \code{list} of length 50.
}
\usage{
SurfTrac
}
\description{
List of element matrices with surface traction. Obtained from function: SurfaceTraction
}
\keyword{datasets}
