% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triMesh.R
\docType{data}
\name{triMesh}
\alias{triMesh}
\title{Meshed coordinate points obtained from function: triangulate0}
\format{
An object of class \code{list} of length 2.
}
\usage{
triMesh
}
\description{
Meshed coordinate points obtained from function: triangulate0
}
\keyword{datasets}
